refactor of PixelateText and Change the CHAR_TIME to 1000
