+++ htextview-library/src/main/java/com/hanks/htextview/animatetext/PixelateText.java
@@ -7,7 +7,6 @@ import android.graphics.Canvas;
 import android.graphics.Color;
 import android.graphics.Matrix;
 import android.graphics.Paint;
-import android.graphics.Rect;
 import android.util.AttributeSet;
 import android.util.DisplayMetrics;
 import android.view.WindowManager;
@@ -15,12 +14,7 @@ import android.view.animation.AccelerateDecelerateInterpolator;
 
 import com.hanks.htextview.HTextView;
 import com.hanks.htextview.animatetext.base.IHText;
-import com.hanks.htextview.util.CharacterDiffResult;
-import com.hanks.htextview.util.CharacterUtils;
-import com.hanks.htextview.util.HLog;
-
-import java.util.ArrayList;
-import java.util.List;
+import com.hanks.htextview.util.MathUtils;
 
 /**
  * 缩放动画
@@ -28,239 +22,28 @@ import java.util.List;
  */
 public class PixelateText implements IHText {
 
-    //缩放系数
-    public final static int SCALE = 8;
-    float progress = 0;
-    Paint paint, oldPaint;
-    float charTime = 2000; // 每个字符动画时间 500ms
-    int mostCount = 20; // 最多10个字符同时动画
-    HTextView mHTextView;
-
-    private float[] gaps = new float[100];
-    private float[] oldGaps = new float[100];
-
-    private DisplayMetrics metrics;
+    private static final int MOST_COUNT = 20; // 最多20个字符同时动画
+    private static final float CHAR_TIME = 1000; // 每个字符动画时间 1000ms
+    private float progress;
     private float textSize;
+    private Paint paint, oldPaint, pixPaint;
+    private HTextView mHTextView;
 
     private CharSequence mText;
     private CharSequence mOldText;
-
-    private List<CharacterDiffResult> differentList = new ArrayList<>();
 
     private float oldStartX = 0;
     private float startX = 0;
     private float startY = 0;
 
-    private Bitmap bitmap;
-    private Matrix matrix;
-    private Paint pixPaint;
-    private Canvas pixCanvas;
+    private Bitmap bitmap = Bitmap.createBitmap(700, 200, Bitmap.Config.ARGB_4444);
+    private Canvas pixCanvas = new Canvas(bitmap);
+    private Matrix matrix = new Matrix();
 
-    public static Bitmap fastBlur(Bitmap sbitmap, float radiusf) {
-        Bitmap bitmap = Bitmap.createScaledBitmap(sbitmap, sbitmap.getWidth() / SCALE, sbitmap.getHeight() / SCALE, false);//先缩放图片，增加模糊速度
-        int radius = (int) radiusf;
-        if (radius < 1) {
-            return (null);
-        }
 
-        int w = bitmap.getWidth();
-        int h = bitmap.getHeight();
-
-        int[] pix = new int[w * h];
-        bitmap.getPixels(pix, 0, w, 0, 0, w, h);
-
-        int wm = w - 1;
-        int hm = h - 1;
-        int wh = w * h;
-        int div = radius + radius + 1;
-
-        int r[] = new int[wh];
-        int g[] = new int[wh];
-        int b[] = new int[wh];
-        int rsum, gsum, bsum, x, y, i, p, yp, yi, yw;
-        int vmin[] = new int[Math.max(w, h)];
-
-        int divsum = (div + 1) >> 1;
-        divsum *= divsum;
-        int dv[] = new int[256 * divsum];
-        for (i = 0; i < 256 * divsum; i++) {
-            dv[i] = (i / divsum);
-        }
-
-        yw = yi = 0;
-
-        int[][] stack = new int[div][3];
-        int stackpointer;
-        int stackstart;
-        int[] sir;
-        int rbs;
-        int r1 = radius + 1;
-        int routsum, goutsum, boutsum;
-        int rinsum, ginsum, binsum;
-
-        for (y = 0; y < h; y++) {
-            rinsum = ginsum = binsum = routsum = goutsum = boutsum = rsum = gsum = bsum = 0;
-            for (i = -radius; i <= radius; i++) {
-                p = pix[yi + Math.min(wm, Math.max(i, 0))];
-                sir = stack[i + radius];
-                sir[0] = (p & 0xff0000) >> 16;
-                sir[1] = (p & 0x00ff00) >> 8;
-                sir[2] = (p & 0x0000ff);
-                rbs = r1 - Math.abs(i);
-                rsum += sir[0] * rbs;
-                gsum += sir[1] * rbs;
-                bsum += sir[2] * rbs;
-                if (i > 0) {
-                    rinsum += sir[0];
-                    ginsum += sir[1];
-                    binsum += sir[2];
-                } else {
-                    routsum += sir[0];
-                    goutsum += sir[1];
-                    boutsum += sir[2];
-                }
-            }
-            stackpointer = radius;
-
-            for (x = 0; x < w; x++) {
-
-                r[yi] = dv[rsum];
-                g[yi] = dv[gsum];
-                b[yi] = dv[bsum];
-
-                rsum -= routsum;
-                gsum -= goutsum;
-                bsum -= boutsum;
-
-                stackstart = stackpointer - radius + div;
-                sir = stack[stackstart % div];
-
-                routsum -= sir[0];
-                goutsum -= sir[1];
-                boutsum -= sir[2];
-
-                if (y == 0) {
-                    vmin[x] = Math.min(x + radius + 1, wm);
-                }
-                p = pix[yw + vmin[x]];
-
-                sir[0] = (p & 0xff0000) >> 16;
-                sir[1] = (p & 0x00ff00) >> 8;
-                sir[2] = (p & 0x0000ff);
-
-                rinsum += sir[0];
-                ginsum += sir[1];
-                binsum += sir[2];
-
-                rsum += rinsum;
-                gsum += ginsum;
-                bsum += binsum;
-
-                stackpointer = (stackpointer + 1) % div;
-                sir = stack[(stackpointer) % div];
-
-                routsum += sir[0];
-                goutsum += sir[1];
-                boutsum += sir[2];
-
-                rinsum -= sir[0];
-                ginsum -= sir[1];
-                binsum -= sir[2];
-
-                yi++;
-            }
-            yw += w;
-        }
-        for (x = 0; x < w; x++) {
-            rinsum = ginsum = binsum = routsum = goutsum = boutsum = rsum = gsum = bsum = 0;
-            yp = -radius * w;
-            for (i = -radius; i <= radius; i++) {
-                yi = Math.max(0, yp) + x;
-
-                sir = stack[i + radius];
-
-                sir[0] = r[yi];
-                sir[1] = g[yi];
-                sir[2] = b[yi];
-
-                rbs = r1 - Math.abs(i);
-
-                rsum += r[yi] * rbs;
-                gsum += g[yi] * rbs;
-                bsum += b[yi] * rbs;
-
-                if (i > 0) {
-                    rinsum += sir[0];
-                    ginsum += sir[1];
-                    binsum += sir[2];
-                } else {
-                    routsum += sir[0];
-                    goutsum += sir[1];
-                    boutsum += sir[2];
-                }
-
-                if (i < hm) {
-                    yp += w;
-                }
-            }
-            yi = x;
-            stackpointer = radius;
-            for (y = 0; y < h; y++) {
-                // Preserve alpha channel: ( 0xff000000 & pix[yi] )
-                pix[yi] = (0xff000000 & pix[yi]) | (dv[rsum] << 16) | (dv[gsum] << 8) | dv[bsum];
-
-                rsum -= routsum;
-                gsum -= goutsum;
-                bsum -= boutsum;
-
-                stackstart = stackpointer - radius + div;
-                sir = stack[stackstart % div];
-
-                routsum -= sir[0];
-                goutsum -= sir[1];
-                boutsum -= sir[2];
-
-                if (x == 0) {
-                    vmin[y] = Math.min(y + r1, hm) * w;
-                }
-                p = x + vmin[y];
-
-                sir[0] = r[p];
-                sir[1] = g[p];
-                sir[2] = b[p];
-
-                rinsum += sir[0];
-                ginsum += sir[1];
-                binsum += sir[2];
-
-                rsum += rinsum;
-                gsum += ginsum;
-                bsum += binsum;
-
-                stackpointer = (stackpointer + 1) % div;
-                sir = stack[stackpointer];
-
-                routsum += sir[0];
-                goutsum += sir[1];
-                boutsum += sir[2];
-
-                rinsum -= sir[0];
-                ginsum -= sir[1];
-                binsum -= sir[2];
-
-                yi += w;
-            }
-        }
-        bitmap.setPixels(pix, 0, w, 0, 0, w, h);
-        return (bitmap);
-
-    }
-
+    @Override
     public void init(HTextView hTextView, AttributeSet attrs, int defStyle) {
         mHTextView = hTextView;
-
-        mText = "";
-        mOldText = "";
 
         paint = new Paint(Paint.ANTI_ALIAS_FLAG);
         paint.setColor(Color.BLACK);
@@ -274,39 +57,29 @@ public class PixelateText implements IHText {
         pixPaint.setColor(Color.BLACK);
         pixPaint.setStyle(Paint.Style.FILL);
 
-        metrics = new DisplayMetrics();
+        DisplayMetrics metrics = new DisplayMetrics();
         WindowManager windowManger = (WindowManager) hTextView.getContext()
                 .getSystemService(Context.WINDOW_SERVICE);
         windowManger.getDefaultDisplay().getMetrics(metrics);
 
         textSize = hTextView.getTextSize();
-
-        bitmap = Bitmap.createBitmap(700, 200, Bitmap.Config.ARGB_4444);
-
-        matrix = new Matrix();
-
-        pixCanvas = new Canvas(bitmap);
     }
 
     @Override
     public void reset(CharSequence text) {
         progress = 1;
-        calc();
+        calculate();
         mHTextView.invalidate();
     }
 
     @Override
     public void animateText(CharSequence text) {
-        mOldText = mText;
+        mOldText = mText == null ? "" : mText;
         mText = text;
-        calc();
-
-        int n = mText.length();
-        n = n <= 0 ? 1 : n;
-
+        calculate();
+        int n = mText.length() < 1 ? 1 : mText.length();
         // 计算动画总时间
-        long duration = (long) (charTime + charTime / mostCount * (n - 1));
-
+        long duration = (long) (CHAR_TIME + CHAR_TIME / MOST_COUNT * (n - 1));
         ValueAnimator valueAnimator = ValueAnimator.ofFloat(0, duration).setDuration(duration);
         valueAnimator.setInterpolator(new AccelerateDecelerateInterpolator());
         valueAnimator.addUpdateListener(new ValueAnimator.AnimatorUpdateListener() {
@@ -323,54 +96,27 @@ public class PixelateText implements IHText {
     public void onDraw(Canvas canvas) {
         float offset = startX;
         float oldOffset = oldStartX;
-
-        float pp = progress / (charTime + charTime / mostCount * (mText.length() - 1));
-
-        HLog.i(pp);
-
-        int alpha = (int) (255 * pp);
-        int oldAlpha = (int) (255 * (1 - pp));
-
-        alpha = alpha > 255 ? 255 : alpha;
-        alpha = alpha < 0 ? 0 : alpha;
-
-        oldAlpha = oldAlpha > 255 ? 255 : oldAlpha;
-        oldAlpha = oldAlpha < 0 ? 0 : oldAlpha;
+        float percent = progress / (CHAR_TIME + CHAR_TIME / MOST_COUNT * mText.length());
+        int alpha = (int) (255 * percent);
+        int oldAlpha = (int) (255 * (1 - percent));
+        alpha = MathUtils.constrain(0, 255, alpha);
+        oldAlpha = MathUtils.constrain(0, 255, oldAlpha);
 
         oldPaint.setAlpha(oldAlpha);
         paint.setAlpha(alpha);
-
         pixCanvas.drawColor(Color.WHITE);
         pixCanvas.drawText(mOldText, 0, mOldText.length(), oldOffset, startY, oldPaint);
         pixCanvas.drawText(mText, 0, mText.length(), offset, startY, paint);
-
-        Rect targetRect = new Rect(0, 0, 700, 200);
         canvas.drawBitmap(bitmap, matrix, pixPaint);
-
     }
 
-    private void calc() {
+    private void calculate() {
         textSize = mHTextView.getTextSize();
         paint.setTextSize(textSize);
-
-        for (int i = 0; i < mText.length(); i++) {
-            gaps[i] = paint.measureText(mText.charAt(i) + "");
-        }
-
         oldPaint.setTextSize(textSize);
-
-        for (int i = 0; i < mOldText.length(); i++) {
-            oldGaps[i] = oldPaint.measureText(mOldText.charAt(i) + "");
-        }
-
         oldStartX = (mHTextView.getWidth() - oldPaint.measureText(mOldText.toString())) / 2f;
-
         startX = (mHTextView.getWidth() - paint.measureText(mText.toString())) / 2f;
         startY = (int) ((mHTextView.getHeight() / 2) - ((paint.descent() + paint.ascent()) / 2));
-
-        differentList.clear();
-        differentList.addAll(CharacterUtils.diff(mOldText, mText));
-
         oldPaint.setTextSize(textSize);
     }
